+++ sc/social/like/subscribers.py
@@ -19,6 +19,10 @@ from sc.social.like.config import IS_PLONE_5
 from sc.social.like.config import PROJECTNAME
 from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.logger import logger
+from sc.social.like.utils import get_content_image
+from sc.social.like.utils import validate_og_description
+from sc.social.like.utils import validate_og_lead_image
+from sc.social.like.utils import validate_og_title
 from zope.component import getUtility
 from zope.schema.interfaces import WrongType
 
@@ -107,3 +111,46 @@ def assign_canonical_url(obj, event):
             'Canonical domain not set in Social Media configlet; '
             "Facebook's Open Graph canonical URL (og:orl) will not be available"
         )
+
+
+def check_sharing_best_practices(obj):
+    """Check if content follows social networks sharing best practices
+    as defined by Twitter and Facebook.
+    """
+    record = ISocialLikeSettings.__identifier__ + '.validation_enabled'
+    validation_enabled = api.portal.get_registry_record(record, default=False)
+    if not validation_enabled:
+        return
+
+    request = obj.REQUEST
+
+    title = getattr(obj, 'title', '')
+    try:
+        validate_og_title(title)
+    except ValueError as e:
+        api.portal.show_message(message=e.message, request=request, type='warning')
+
+    description = getattr(obj, 'description', '')
+    try:
+        validate_og_description(description)
+    except ValueError as e:
+        api.portal.show_message(message=e.message, request=request, type='warning')
+
+    image = get_content_image(obj)
+    try:
+        validate_og_lead_image(image)
+    except ValueError as e:
+        api.portal.show_message(message=e.message, request=request, type='warning')
+
+
+def check_sharing_best_practices_on_editing(obj, event):
+    """Event subscriber for content being edited."""
+    state = api.content.get_state(obj)
+    if state in ('published', ):
+        check_sharing_best_practices(obj)
+
+
+def check_sharing_best_practices_on_publishing(obj, event):
+    """Event subscriber for content being published."""
+    if event.status['review_state'] in ('published', ):
+        check_sharing_best_practices(obj)