+++ sc/social/like/tests/test_subscribers.py
@@ -1,14 +1,29 @@
 # -*- coding: utf-8 -*-
+"""Subscribers related test cases.
+
+Sharing best practices validation tests run only with Dexterity-based
+content types.
+"""
+from plone import api
+from plone.dexterity.events import EditFinishedEvent
 from plone.registry import field
 from plone.registry.interfaces import IRegistry
 from plone.registry.record import Record
 from plone.supermodel import model
+from Products.statusmessages.interfaces import IStatusMessage
 from sc.social.like.config import IS_PLONE_5
 from sc.social.like.config import PROJECTNAME
 from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.testing import INTEGRATION_TESTING
+from sc.social.like.testing import load_image
+from sc.social.like.tests.api_hacks import set_image_field
+from sc.social.like.tests.utils import get_random_string
+from sc.social.like.utils import MSG_INVALID_OG_DESCRIPTION
+from sc.social.like.utils import MSG_INVALID_OG_LEAD_IMAGE_DIMENSIONS
+from sc.social.like.utils import MSG_INVALID_OG_TITLE
 from zope import schema
 from zope.component import getUtility
+from zope.event import notify
 
 import unittest
 
@@ -69,3 +84,133 @@ class ControlPanelTestCase(unittest.TestCase):
         # both records must be unchanged
         self.assertEqual(self.settings.twitter_username, '')
         self.assertEqual(self.registry['plone.twitter_username'], '')
+
+
+class ValidationTestCase(unittest.TestCase):
+
+    layer = INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+
+        with api.env.adopt_roles(['Manager']):
+            self.news_item = api.content.create(
+                self.portal, 'News Item', title=u'Lorem ipsum')
+        set_image_field(self.news_item, load_image(1024, 768), 'image/png')
+
+    def test_do_not_validate_on_submit(self):
+        # content don't follow best practices
+        self.news_item.title = get_random_string(80)
+        self.news_item.description = get_random_string(300)
+        set_image_field(self.news_item, load_image(200, 150), 'image/png')
+
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'submit')
+
+        # no feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 0)
+
+    def test_validate_on_publish_valid(self):
+        # content follows best practices
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'publish')
+
+        # no feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 0)
+
+    def test_validate_on_publish_invalid(self):
+        # content don't follow best practices
+        self.news_item.title = get_random_string(80)
+        self.news_item.description = get_random_string(300)
+        set_image_field(self.news_item, load_image(200, 150), 'image/png')
+
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'publish')
+
+        # feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 3)
+        self.assertEqual(messages[0].message, MSG_INVALID_OG_TITLE)
+        self.assertEqual(messages[0].type, u'warning')
+        self.assertEqual(messages[1].message, MSG_INVALID_OG_DESCRIPTION)
+        self.assertEqual(messages[1].type, u'warning')
+        self.assertEqual(messages[2].message, MSG_INVALID_OG_LEAD_IMAGE_DIMENSIONS)
+        self.assertEqual(messages[2].type, u'warning')
+
+    def test_do_not_validate_on_edit_not_public(self):
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'submit')
+
+        # content don't follow best practices
+        self.news_item.title = get_random_string(80)
+        self.news_item.description = get_random_string(300)
+        set_image_field(self.news_item, load_image(200, 150), 'image/png')
+        notify(EditFinishedEvent(self.news_item))
+
+        # no feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 0)
+
+    def test_validate_on_edit_valid(self):
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'publish')
+
+        # content follows best practices
+        self.news_item.title = get_random_string(70)
+        self.news_item.description = get_random_string(200)
+        set_image_field(self.news_item, load_image(1920, 1080), 'image/png')
+        notify(EditFinishedEvent(self.news_item))
+
+        # no feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 0)
+
+    def test_validate_social_content_edit_invalid(self):
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'publish')
+
+        # content don't follow best practices
+        self.news_item.title = get_random_string(80)
+        self.news_item.description = get_random_string(300)
+        set_image_field(self.news_item, load_image(200, 150), 'image/png')
+        notify(EditFinishedEvent(self.news_item))
+
+        # no feedback messages present
+        messages = IStatusMessage(self.request).show()
+        self.assertEqual(len(messages), 3)
+        self.assertEqual(messages[0].message, MSG_INVALID_OG_TITLE)
+        self.assertEqual(messages[0].type, u'warning')
+        self.assertEqual(messages[1].message, MSG_INVALID_OG_DESCRIPTION)
+        self.assertEqual(messages[1].type, u'warning')
+        self.assertEqual(messages[2].message, MSG_INVALID_OG_LEAD_IMAGE_DIMENSIONS)
+        self.assertEqual(messages[2].type, u'warning')
+
+    def test_validate_with_package_uninstalled(self):
+        from sc.social.like.config import PROJECTNAME
+        qi = self.portal['portal_quickinstaller']
+        qi.uninstallProducts(products=[PROJECTNAME])
+
+        # should not raise exceptions on publishing
+        with api.env.adopt_roles(['Manager']):
+            api.content.transition(self.news_item, 'publish')
+
+        # should not raise exceptions on editing
+        notify(EditFinishedEvent(self.news_item))
+
+
+def load_tests(loader, tests, pattern):
+    from sc.social.like.testing import HAS_DEXTERITY
+
+    test_cases = [ControlPanelTestCase]
+    if HAS_DEXTERITY:
+        # load validation tests on Dexterity-based content types only
+        test_cases.append(ValidationTestCase)
+
+    suite = unittest.TestSuite()
+    for test_class in test_cases:
+        tests = loader.loadTestsFromTestCase(test_class)
+        suite.addTests(tests)
+    return suite