+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.11b3 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Implement a validator to check if content follows social networks sharing best practices (closes `#128 <https://github.com/collective/sc.social.like/issues/128>`_).
+  [claytonc, hvelarde]
 
 
 2.11b2 (2017-09-11)